Move common body logic into _BodyBase
