+++ planetmapper/body.py
@@ -20,7 +20,7 @@ from spiceypy.utils.exceptions import (
 )
 
 from . import data_loader, utils
-from .base import Numeric, SpiceBase
+from .base import Numeric, _BodyBase
 from .basic_body import BasicBody
 
 _WireframeComponent = Literal[
@@ -89,7 +89,7 @@ DEFAULT_WIREFRAME_FORMATTING: dict[_WireframeComponent, dict[str, Any]] = {
 }
 
 
-class Body(SpiceBase):
+class Body(_BodyBase):
     """
     Class representing an astronomical body observed at a specific time.
 
@@ -144,7 +144,14 @@ class Body(SpiceBase):
         surface_method: str = 'ELLIPSOID',
         **kwargs,
     ) -> None:
-        super().__init__(**kwargs)
+        super().__init__(
+            target=target,
+            utc=utc,
+            observer=observer,
+            aberration_correction=aberration_correction,
+            observer_frame=observer_frame,
+            **kwargs,
+        )
 
         # Document instance variables
         self.et: float
@@ -254,39 +261,15 @@ class Body(SpiceBase):
         """
 
         # Process inputs
-        if isinstance(utc, float):
-            utc = self.mjd2dtm(utc)
-        if utc is None:
-            utc = datetime.datetime.now(datetime.timezone.utc)
-        if isinstance(utc, datetime.datetime):
-            # convert input datetime to UTC, then to a string compatible with spice
-            utc = utc.replace(tzinfo=datetime.timezone.utc)
-            utc = utc.strftime(self._DEFAULT_DTM_FORMAT_STRING)
-
-        self.target = self.standardise_body_name(target)
-        self.observer = self.standardise_body_name(observer)
-        self.observer_frame = observer_frame
         self.illumination_source = illumination_source
-        self.aberration_correction = aberration_correction
         self.subpoint_method = subpoint_method
         self.surface_method = surface_method
 
-        # Encode strings which are regularly passed to spice (for speed)
-        self._target_encoded = self._encode_str(self.target)
-        self._observer_encoded = self._encode_str(self.observer)
-        self._observer_frame_encoded = self._encode_str(self.observer_frame)
         self._illumination_source_encoded = self._encode_str(self.illumination_source)
-        self._aberration_correction_encoded = self._encode_str(
-            self.aberration_correction
-        )
         self._subpoint_method_encoded = self._encode_str(self.subpoint_method)
         self._surface_method_encoded = self._encode_str(self.surface_method)
 
         # Get target properties and state
-        self.et = spice.utc2et(utc)
-        self.dtm: datetime.datetime = self.et2dtm(self.et)
-        self.utc = self.dtm.strftime(self._DEFAULT_DTM_FORMAT_STRING)
-        self.target_body_id: int = spice.bodn2c(self.target)
         self.target_frame = 'IAU_' + self.target
         self._target_frame_encoded = self._encode_str(self.target_frame)
 
@@ -308,23 +291,6 @@ class Body(SpiceBase):
         else:
             self.positive_longitude_direction = 'E'
 
-        starg, lt = spice.spkezr(
-            self._target_encoded,  # type: ignore
-            self.et,
-            self._observer_frame_encoded,  # type: ignore
-            self._aberration_correction_encoded,  # type: ignore
-            self._observer_encoded,  # type: ignore
-        )
-        self._target_obsvec = cast(np.ndarray, starg)[:3]
-        self.target_light_time = cast(float, lt)
-        # cast() calls are only here to make type checking play nicely with spice.spkezr
-        self.target_distance = self.target_light_time * self.speed_of_light()
-        self._target_ra_radians, self._target_dec_radians = self._obsvec2radec_radians(
-            self._target_obsvec
-        )
-        self.target_ra, self.target_dec = self._radian_pair2degrees(
-            self._target_ra_radians, self._target_dec_radians
-        )
         self.target_diameter_arcsec = (
             60 * 60 * np.rad2deg(np.arcsin(2 * self.r_eq / self.target_distance))
         )
@@ -384,12 +350,7 @@ class Body(SpiceBase):
 
     def _get_equality_tuple(self) -> tuple:
         return (
-            self.target,
-            self.utc,
-            self.observer,
-            self.observer_frame,
             self.illumination_source,
-            self.aberration_correction,
             self.subpoint_method,
             self.surface_method,
             super()._get_equality_tuple(),
@@ -614,13 +575,6 @@ class Body(SpiceBase):
             self.et,
         )
         return np.matmul(px, rayvec)
-
-    def _obsvec2radec_radians(self, obsvec: np.ndarray) -> tuple[float, float]:
-        """
-        Transform rectangular vector in observer frame to observer ra/dec coordinates.
-        """
-        dst, ra, dec = spice.recrad(obsvec)
-        return ra, dec
 
     # Coordinate transformations observer -> target direction
     def _radec2obsvec_norm_radians(self, ra: float, dec: float) -> np.ndarray:
